Update to Sentiment.php
Score and categorise functions have been modified.  Score now decides an overall dominant sentiment (in case of a tie in scores), which is added to the $scores array as key "dom".  Categorise retrieves this winning sentiment ("dom") rather than the first key of the $scores array as assumed winning sentiment.